+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added `api.ip` and `api.secure` as server options
 -   API endpoints for adding/removing ip allow/deny list entries
 -   Added tests for stager compilation using `DotnetCompiler`
+-   Added `-f` flag for install script to force install as root
 -   Added dynamic options to modules
 -   Added PT, TCH rules to ruff config
 
@@ -68,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Removed dotnet install in Docker
 -   Removed `nim` stager from Empire
 -   Removed nim from installation script
+-   Removed migration docs to Empire 5.0
 
 #### Breaking